+++ changelog.md
@@ -1,1 +1,1 @@
-* Turkish localization by Umut KORKMAZ 
+* The category jump gesture is now generated using vkCode instead of ASCII. This way it should work on keyboards with non-latin alphabets.